% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_bloco}
\alias{peg_bloco}
\title{Pegar informações de um ou mais blocos de partidos}
\usage{
peg_bloco(id_bloco)
}
\arguments{
\item{id_bloco}{Identificador(es) do(s) bloco(s)}
}
\description{
Pegar informações de um ou mais blocos de partidos
}
