% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_deputado}
\alias{peg_deputado}
\title{Pegar informações de um ou mais deputados}
\usage{
peg_deputado(id_deputado)
}
\arguments{
\item{id_deputado}{Identificador do deputado}
}
\description{
Pegar informações de um ou mais deputados
}
