% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_evento}
\alias{peg_evento}
\title{Pegar informações de um ou mais eventos}
\usage{
peg_evento(id_evento)
}
\arguments{
\item{id_evento}{Identificador do evento}
}
\description{
Pegar informações de um ou mais eventos
}
