% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_legislatura}
\alias{peg_legislatura}
\title{Pegar informações de uma ou mais legislaturas}
\usage{
peg_legislatura(id_legislatura)
}
\arguments{
\item{id_legislatura}{Identificador(es) da(s) legislatura(s)}
}
\description{
Pegar informações de uma ou mais legislaturas
}
