% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_mesa}
\alias{peg_mesa}
\title{Pegar informações de uma ou mais mesas de legislaturas}
\usage{
peg_mesa(id_legislatura)
}
\arguments{
\item{id_legislatura}{Identificador(es) da(s) legislatura(s)}
}
\description{
Pegar informações de uma ou mais mesas de legislaturas
}
