% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_orgao}
\alias{peg_orgao}
\title{Pegar informações de um ou mais órgãos legislativos}
\usage{
peg_orgao(id_orgao)
}
\arguments{
\item{id_orgao}{Identificador do órgão}
}
\description{
Pegar informações de um ou mais órgãos legislativos
}
