% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_referencia}
\alias{peg_referencia}
\title{Pegar tabela de referência de alguma categoria}
\usage{
peg_referencia(tipo_referencia)
}
\arguments{
\item{tipo_referencia}{Tipo da referência (situacoesDeputado,
situacoesEvento, situacoesOrgao, situacoesProposicao, tiposEvento,
tiposOrgao, tiposProposicao, uf)}
}
\description{
Pegar tabela de referência de alguma categoria
}
