% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_blocos}
\alias{pesq_blocos}
\title{Pesquisar blocos de partidos}
\usage{
pesq_blocos(id_legislatura = NULL, sigla_partido = NULL, n_max = 15)
}
\arguments{
\item{id_legislatura}{Identificador(es) da(s) legislatura(s) para filtrar}

\item{sigla_partido}{Sigla(s) do(s) partido(s) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar blocos de partidos
}
