% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_deputados}
\alias{pesq_deputados}
\title{Pesquisar deputados}
\usage{
pesq_deputados(id_legislatura = NULL, sigla_uf = NULL,
  sigla_partido = NULL, sigla_sexo = NULL, n_max = 15)
}
\arguments{
\item{id_legislatura}{Identificador(es) da(s) legislatura(s) para filtrar}

\item{sigla_uf}{Unidade(s) federativa(s) para filtrar}

\item{sigla_partido}{Sigla(s) do(s) partido(s) para filtrar}

\item{sigla_sexo}{Gênero(s) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar deputados
}
