% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_despesas_deputado}
\alias{pesq_despesas_deputado}
\title{Pesquisar despesas de um ou mais deputados}
\usage{
pesq_despesas_deputado(id_deputado, id_legislatura = NULL, ano = NULL,
  mes = NULL, cnpj_cpf_fornecedor = NULL, n_max = 15)
}
\arguments{
\item{id_deputado}{Identificador(es) do(s) deputado(s) para filtrar}

\item{id_legislatura}{Identificador(es) da(s) legislatura(s) para filtrar}

\item{ano}{Ano(s) para filtrar}

\item{mes}{Mês(es) para filtrar}

\item{cnpj_cpf_fornecedor}{CNPJ/CPF do(s) fornecedor(es) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar despesas de um ou mais deputados
}
