% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_eventos}
\alias{pesq_eventos}
\title{Pesquisar eventos ocorridos ou previstos nos diversos órgãos da Câmara}
\usage{
pesq_eventos(id_tipo_evento = NULL, id_situacao = NULL,
  id_tipo_orgao = NULL, id_orgao = NULL, data_inicio = NULL,
  data_fim = NULL, hora_inicio = NULL, hora_fim = NULL, n_max = 15)
}
\arguments{
\item{id_tipo_evento}{Identificador(es) do(s) tipo(s) de evento para filtrar}

\item{id_situacao}{Identificador(es) da(s) situação(ões) para filtrar}

\item{id_tipo_orgao}{Identificador(es) do(s) tipo(s) de órgão para filtrar}

\item{id_orgao}{Identificador(es) do(s) órgão(s) para filtrar}

\item{data_inicio}{Data(s) de início para filtrar}

\item{data_fim}{Data(s) de término para filtrar}

\item{hora_inicio}{Hora(s) de início para filtrar}

\item{hora_fim}{Hora(s) de término para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar eventos ocorridos ou previstos nos diversos órgãos da Câmara
}
