% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_eventos_deputado}
\alias{pesq_eventos_deputado}
\title{Pesquisar eventos de um ou mais deputados}
\usage{
pesq_eventos_deputado(id_deputado, data_inicio = NULL, data_fim = NULL,
  n_max = 15)
}
\arguments{
\item{id_deputado}{Identificador(es) do(s) deputado(s) para filtrar}

\item{data_inicio}{Data(s) de início para filtrar}

\item{data_fim}{Data(s) de término para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar eventos de um ou mais deputados
}
