% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_eventos_orgao}
\alias{pesq_eventos_orgao}
\title{Pesquisar eventos de um órgão}
\usage{
pesq_eventos_orgao(id_orgao, id_tipo_evento = NULL, n_max = 15)
}
\arguments{
\item{id_orgao}{Identificador(es) do(s) órgão(s) para filtrar}

\item{id_tipo_evento}{Identificador(es) do(s) tipo(s) de evento para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar eventos de um órgão
}
