% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_legislaturas}
\alias{pesq_legislaturas}
\title{Pesquisar legislaturas}
\usage{
pesq_legislaturas(data = NULL, n_max = 15)
}
\arguments{
\item{data}{Data(s) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar legislaturas
}
