% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_orgaos}
\alias{pesq_orgaos}
\title{Pesquisar órgãos legislativos}
\usage{
pesq_orgaos(id_tipo_orgao = NULL, data_inicio = NULL, data_fim = NULL,
  n_max = 15)
}
\arguments{
\item{id_tipo_orgao}{Identificador(es) do(s) tipo(s) de órgão para filtrar}

\item{data_inicio}{Data(s) de início para filtrar}

\item{data_fim}{Data(s) de término para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar órgãos legislativos
}
