% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_partidos}
\alias{pesq_partidos}
\title{Pesquisar partidos}
\usage{
pesq_partidos(data_inicio = NULL, data_fim = NULL, id_legislatura = NULL,
  n_max = 15)
}
\arguments{
\item{data_inicio}{Data(s) de início para filtrar}

\item{data_fim}{Data(s) de término para filtrar}

\item{id_legislatura}{Identificador(es) da(s) legislatura(s) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar partidos
}
