% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_proposicoes}
\alias{pesq_proposicoes}
\title{Pesquisar proposições}
\usage{
pesq_proposicoes(sigla_uf_autor = NULL, sigla_tipo = NULL,
  sigla_partido_autor = NULL, numero = NULL, ano = NULL,
  data_aprensentacao_inicio = NULL, data_aprensentacao_fim = NULL,
  data_inicio = NULL, data_fim = NULL, id_autor = NULL,
  nome_autor = NULL, id_situacao = NULL, cod_partido = NULL, n_max = 15)
}
\arguments{
\item{sigla_uf_autor}{Unidade(s) federativa(s) do(s) autor(es) para filtrar}

\item{sigla_tipo}{Sigla(s) do(s) tipo(s) de proposição(ões) para filtrar}

\item{sigla_partido_autor}{Sigla(s) do(s) partido(s) do(s) autor(es) para filtrar}

\item{numero}{Número(s) da(s) proposição(ões) para filtrar}

\item{ano}{Ano(s) para filtrar}

\item{data_aprensentacao_inicio}{Data(s) de início da(s) apresentação(ções) para filtrar}

\item{data_aprensentacao_fim}{Data(s) de término da(s) apresentação(ções) para filtrar}

\item{data_inicio}{Data(s) de início para filtrar}

\item{data_fim}{Data(s) de término para filtrar}

\item{id_autor}{Identificador(es) do(s) autor(es) para filtrar}

\item{nome_autor}{Nome(s) do(s) autor(es) para filtrar}

\item{id_situacao}{Identificador(es) da(s) situação(ções) para filtrar}

\item{cod_partido}{Código(s) do(s) partido(s) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar proposições
}
