% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesq.R
\name{pesq_tramitacoes_proposicao}
\alias{pesq_tramitacoes_proposicao}
\title{Pesquisar tramitações de uma proposição}
\usage{
pesq_tramitacoes_proposicao(id_proposicao, n_max = 15)
}
\arguments{
\item{id_proposicao}{Identificador(es) da(s) proposição(ções) para filtrar}

\item{n_max}{Número máximo de registros para retornar}
}
\description{
Pesquisar tramitações de uma proposição
}
