% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacoes_deputado}
\alias{situacoes_deputado}
\title{Situações em que um deputado pode se encontrar e suas abreviações}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.}
\usage{
situacoes_deputado
}
\description{
Situações em que um deputado pode se encontrar e suas abreviações
}
\keyword{datasets}
