% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacoes_evento}
\alias{situacoes_evento}
\title{Situações em que um evento pode se encontrar e seus identificadores}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.}
\usage{
situacoes_evento
}
\description{
Situações em que um evento pode se encontrar e seus identificadores
}
\keyword{datasets}
