% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacoes_orgao}
\alias{situacoes_orgao}
\title{Situações em que um orgão pode se encontrar e seus identificadores}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.}
\usage{
situacoes_orgao
}
\description{
Situações em que um orgão pode se encontrar e seus identificadores
}
\keyword{datasets}
