% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacoes_proposicao}
\alias{situacoes_proposicao}
\title{Situações em que uma proposição pode se encontrar e seus identificadores}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 79 rows and 2 columns.}
\usage{
situacoes_proposicao
}
\description{
Situações em que uma proposição pode se encontrar e seus identificadores
}
\keyword{datasets}
