% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipos_evento}
\alias{tipos_evento}
\title{Tipos de evento e seus identificadores}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 2 columns.}
\usage{
tipos_evento
}
\description{
Tipos de evento e seus identificadores
}
\keyword{datasets}
