% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipos_orgao}
\alias{tipos_orgao}
\title{Tipos de orgão e suas abreviações}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 2 columns.}
\usage{
tipos_orgao
}
\description{
Tipos de orgão e suas abreviações
}
\keyword{datasets}
