% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tipos_proposicao}
\alias{tipos_proposicao}
\title{Tipos de proposição, seus identificadores e suas abreviações}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 419 rows and 4 columns.}
\usage{
tipos_proposicao
}
\description{
Tipos de proposição, seus identificadores e suas abreviações
}
\keyword{datasets}
