% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uf}
\alias{uf}
\title{Nomes dos estados, seus identificadores e suas abreviações}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 3 columns.}
\usage{
uf
}
\description{
Nomes dos estados, seus identificadores e suas abreviações
}
\keyword{datasets}
