% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_partido}
\alias{peg_partido}
\title{Pegar informações de um ou mais partidos}
\usage{
peg_partido(id_partido)
}
\arguments{
\item{id_partido}{Identificador do partido}
}
\description{
Pegar informações de um ou mais partidos
}
