% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg.R
\name{peg_proposicao}
\alias{peg_proposicao}
\title{Pegar informações de uma ou mais proposições}
\usage{
peg_proposicao(id_proposicao)
}
\arguments{
\item{id_proposicao}{Identificador(es) da(s) proposição(s)}
}
\description{
Pegar informações de uma ou mais proposições
}
